+++ test/sqlancer/postgres/PostgresBombardTest.java
@@ -42,7 +42,9 @@ class PostgresBombardTest {
         assertTrue(PostgresProvider.isExcludedBombardAction(PostgresProvider.Action.CLUSTER));
         assertTrue(PostgresProvider.isExcludedBombardAction(PostgresProvider.Action.REINDEX));
         assertTrue(PostgresProvider.isExcludedBombardAction(PostgresProvider.Action.CREATE_PARTITION));
+        assertTrue(PostgresProvider.isExcludedBombardAction(PostgresProvider.Action.ATTACH_PARTITION));
         assertTrue(PostgresProvider.isExcludedBombardAction(PostgresProvider.Action.DETACH_PARTITION));
+        assertTrue(PostgresProvider.isExcludedBombardAction(PostgresProvider.Action.DROP_PARTITION));
 
         assertFalse(PostgresProvider.isExcludedBombardAction(PostgresProvider.Action.INSERT));
         assertFalse(PostgresProvider.isExcludedBombardAction(PostgresProvider.Action.UPDATE));